==> MD_code/01_Annotation.Rmd <==
---
title: "Annotation Setup"
author: "Lucas Rizkalla, Mikhail Dozmorov"
date: "`r Sys.Date()`"
output:
  html_document:
    theme: cerulean
    toc: yes
  pdf_document:
    toc: no
---

```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
# Set up the environment
library(knitr)
opts_chunk$set(cache.path='cache/', fig.path='img/', cache=F, tidy=T, fig.keep='high', echo=F, dpi=100, warnings=F, message=F, comment=NA, warning=F, results='as.is', fig.width = 10, fig.height = 6) #out.width=700, 
library(pander)
panderOptions('table.split.table', Inf)
set.seed(1)
library(dplyr)
options(stringsAsFactors = FALSE)
```

```{r libraries}
library(readxl)
library(tidyverse)
```

```{r settings}
# Lucas's path
data.dir           = "/Users/lucasrizkalla/Desktop/Methylation Research-VCU/Methylation850K/data"
# Mikhail's path
data.dir           = "/Users/mdozmorov/Documents/Work/GitHub/Methylation850K/data"
# Annotation files
sample.annotation  = file.path(data.dir, "FRAPData07Feb2020-Mikhail.xlsx")
description2sample = file.path(data.dir, "library_ids_haib20TS6106.csv")
sample2idat        = file.path(data.dir, "Sample_Table.csv")
# Output file
fileNameOut1       <- file.path(data.dir, "Sample_Sheet.csv")
```

# Fill out EXCEL File with remaining ages and race 
```{r}
myAnnotation = read_xlsx(sample.annotation)
# Append missing "030-B"
myAnnotation <- rbind(myAnnotation, c("030-B", "030", "B", rep(NA, ncol(myAnnotation) - 3)))
myAnnotation <- myAnnotation[order(myAnnotation$Identifier), ]

IDCount = 1
for (ID in myAnnotation$ID) { # Go through unique IDs only
  if ((myAnnotation$ID[IDCount] == myAnnotation$ID[IDCount + 1]) & (IDCount < nrow(myAnnotation))) {
    myAnnotation$Age[IDCount + 1] <- myAnnotation$Age[IDCount]
    myAnnotation$Race[IDCount + 1] <- myAnnotation$Race[IDCount]
  }
  IDCount = IDCount + 1
}
```

# Creation of a Sample Annotation file that matches IDs of Idat files

```{r}
myIdatName   = read.csv(sample2idat)
myIdentifier = read.csv(description2sample, as.is = TRUE)

# Subset to common IDs
setdiff(myIdentifier$Description, myAnnotation$Identifier)
common_id    <- intersect(myAnnotation$Identifier, myIdentifier$Description)
myIdentifier <- myIdentifier[myIdentifier$Description %in% common_id, ]
myIdatName   <- myIdatName[myIdatName$Sample %in% myIdentifier$Sample.Name, ]

#Create new column to be used as idat file name (SentrixID_SentrixPosition)
myIdatName$Sentrix_ID = myIdatName$BeadChip
myIdatName$Sentrix_Position = paste(myIdatName$SSR, myIdatName$SSC, sep = "")

#For use in RnBeads
myIdatName$barcode = paste(myIdatName$BeadChip, "_", myIdatName$SSR, myIdatName$SSC, sep = "")

#Sort by SampleName in order to cross reference with Sample in myIdatName (should be identical in order)
myIdentifier <- myIdentifier[match(myIdatName$Sample, myIdentifier$Sample.Name), ]
all.equal(myIdentifier$Sample.Name, myIdatName$Sample) # TRUE if order matches
#Replace column with corresponding IDs that will be found in myAnnotation 
myIdatName$Sample <- myIdentifier$Description

#Sort in numerical order so that it correspond correctly to myAnnotation
myIdatName = myIdatName[match(myAnnotation$Identifier, myIdatName$Sample), ] # Works always
all.equal(myIdatName$Sample, myAnnotation$Identifier) # TRUE if the same order

#Once sorted, replaced with correct Idat file name
myAnnotation$Identifier <- myIdatName$Sentrix_ID

#Set column name to "barcode" for ChAMP
names(myAnnotation)[names(myAnnotation) == "Identifier"] <- "Sentrix_ID"
myAnnotation <- myAnnotation %>% add_column("Sentrix_Position" = myIdatName$Sentrix_Position, .after = "Sentrix_ID")
myAnnotation <- myAnnotation %>% add_column("barcode" = myIdatName$barcode, .after = "Sentrix_Position")

myAnnotation <- myAnnotation[order(myAnnotation$Sentrix_ID, myAnnotation$Sentrix_Position),]

mySampleAnnotation <- write.csv(myAnnotation, file = file.path(data.dir, "Sample_Annotation.csv"))
```

# Remove Unpaired Data

```{r}
sample_table <- table(myAnnotation$ID) # Count saples per ID
sample_unpaired <- names(sample_table)[sample_table == 1] # Samples with single ID
sample_unpaired
myAnnotation <- myAnnotation[!(myAnnotation$ID %in% sample_unpaired),  ]
```

# Save the data

```{r}
write.csv(myAnnotation, file = fileNameOut1, row.names = FALSE)
```
